Rename modules and separate models and views.
